requirements.txt
flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
openpyxl==3.1.5
nltk==3.9.1
gunicorn==21.2.0
python-docx==1.1.2
google-generativeai==0.7.2
googlesearch-python==1.3.0
serpapi
tavily-python
python-dotenv